refactor(library): extract listing views into their own file
index.tsx had grown to 1131 lines mixing route/state orchestration with
the pure-presentational listing views (RootView, SearchResultsView,
PublicSetsView, VolumeView) and their shared helpers. None of these read
any local state from LibraryPage/Library — they only take props and call
their own hooks — so they move to library-views.tsx unchanged.
